LayoutTests/imported/w3c/web-platform-tests/custom-elements/custom-element-registry/define-expected.txt

CONSOLE MESSAGE: line 15: TypeError: undefined is not an object (evaluating 'iframe.contentDocument.defaultView')

FAIL Custom Elements: Element definition TypeError: undefined is not an object (evaluating 'iframe.contentDocument.defaultView')
